+++ design-toolkit/icons/tsup.config.ts
@@ -10,21 +10,26 @@
  * governing permissions and limitations under the License.
  */
 
-import { esbuildPluginFilePathExtensions } from 'esbuild-plugin-file-path-extensions';
+import {
+  fixAliasPlugin,
+  fixExtensionsPlugin,
+  fixFolderImportsPlugin,
+} from 'esbuild-fix-imports-plugin';
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
   esbuildPlugins: [
-    esbuildPluginFilePathExtensions({
-      esmExtension: 'js',
-    }),
+    fixAliasPlugin(),
+    fixFolderImportsPlugin(),
+    fixExtensionsPlugin(),
   ],
   entry: ['src/icons/**/*.{ts,tsx}'],
-  bundle: true,
+  bundle: false,
   clean: true,
   dts: true,
   format: 'esm',
   sourcemap: true,
   splitting: true,
   treeshake: true,
+  metafile: true,
 });